- add BaseUiTest.kt to all ui tests
