+++ app/src/androidTest/java/com/fazemeright/myinventorytracker/ui/addbag/AddBagActivityTest.kt
@@ -1,37 +1,23 @@
 package com.fazemeright.myinventorytracker.ui.addbag
 
-import androidx.test.core.app.ActivityScenario
+import androidx.appcompat.app.AppCompatActivity
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.*
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.matcher.ViewMatchers.*
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.fazemeright.myinventorytracker.R
 import com.fazemeright.myinventorytracker.isViewDisplayed
 import com.fazemeright.myinventorytracker.isViewWithTextDisplayed
-import org.junit.After
-import org.junit.Before
+import com.fazemeright.myinventorytracker.ui.base.BaseUiTest
 import org.junit.Test
-import org.junit.runner.RunWith
 
+class AddBagActivityTest : BaseUiTest() {
+    override fun getActivity(): Class<AppCompatActivity> =
+        AddBagActivity::class.java as Class<AppCompatActivity>
 
-@RunWith(AndroidJUnit4::class)
-class AddBagActivityTest {
-
-    @Before
-    fun setUp() {
-        Intents.init()
-        ActivityScenario.launch(AddBagActivity::class.java)
-    }
-
-    @After
-    fun tearDown() {
-        Intents.release()
-    }
 
     @Test
-    fun allViewsAreDisplayed() {
+    override fun allViewsAreDisplayed() {
         R.string.add_bag_title.isViewWithTextDisplayed()
         R.id.edtBagName.isViewDisplayed()
         R.id.btnChooseColor.isViewDisplayed()
