+++ app/src/androidTest/java/com/fazemeright/myinventorytracker/ui/settings/SettingsActivityTest.kt
@@ -1,36 +1,24 @@
 package com.fazemeright.myinventorytracker.ui.settings
 
-import androidx.test.core.app.ActivityScenario
+import androidx.appcompat.app.AppCompatActivity
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.assertion.ViewAssertions.matches
-import androidx.test.espresso.intent.Intents
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withContentDescription
-import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.fazemeright.myinventorytracker.R
 import com.fazemeright.myinventorytracker.isViewDisplayed
 import com.fazemeright.myinventorytracker.isViewWithTextDisplayed
-import org.junit.After
-import org.junit.Before
+import com.fazemeright.myinventorytracker.ui.base.BaseUiTest
 import org.junit.Test
-import org.junit.runner.RunWith
 
-@RunWith(AndroidJUnit4::class)
-class SettingsActivityTest {
+class SettingsActivityTest : BaseUiTest() {
 
-    @Before
-    fun setUp() {
-        Intents.init()
-        ActivityScenario.launch(SettingsActivity::class.java)
-    }
-
-    @After
-    fun tearDown() {
-        Intents.release()
+    override fun getActivity(): Class<AppCompatActivity> {
+        return SettingsActivity::class.java as Class<AppCompatActivity>
     }
 
     @Test
-    fun allViewsAreDisplayed() {
+    override fun allViewsAreDisplayed() {
         R.id.tvCopyright.isViewDisplayed()
         R.id.tvMadeWith.isViewDisplayed()
         R.id.settings.isViewDisplayed()
